Add files via upload
The plugin gives you variety of options to have a specific channel, role being set and removed. Allow for the Minecraft and Discord names to be synced. Configure the code's timeout, length and a lot more. The plugin is highly configurable, extensions can be requested and bugs can be fixed at my discord:
https://discord.gg/dt996BNkXt